src/exported_module/requirements.txt
attrs==21.2.0
pynamodb==5.3.2
boto3==1.26.80
botocore==1.29.80
certifi==2021.5.30
chardet==4.0.0
PyJWT==2.8.0
colorama==0.4.1
configobj==5.0.6
idna==2.10
jsonpickle==1.3
jsonschema==3.2.0
pyrsistent==0.17.3
python-dateutil==2.8.2
regex==2021.3.17
requests==2.31.0
six==1.16.0
tqdm==4.19.5
zipp==3.4.1
dynamodb-json==1.3
pymongo==4.5.0
bcrypt==4.0.1
boltons==23.0.0
bottle==0.12.19
hvac==1.2.1
psutil==5.8.0
XlsxWriter==3.0.3
typing_extensions==4.8.0
pytablewriter==0.64.2  # currently not used
ruamel.yaml==0.17.16
gunicorn==20.1.0
pycryptodome==3.19.0
APScheduler==3.10.4
pydantic==1.10.2
cachetools==5.3.1
python-dotenv==1.0.0
dacite==1.8.1
et-xmlfile==1.1.0
openpyxl==3.0.10
cryptography==3.4.7
modular-sdk==3.3.2